CR-SO-107 Refactoring code after CR.
