+++ lib/presentation/home_page/home_page.dart
@@ -1,11 +1,10 @@
-import 'package:bordered_text/bordered_text.dart';
 import 'package:flutter/material.dart';
 import 'package:hive_flutter/adapters.dart';
 import 'package:icon_decoration/icon_decoration.dart';
 import 'package:weather/const/hive_box_names.dart';
 import 'package:weather/presentation/search_page/search_page.dart';
 import 'package:weather/presentation/settings_page/settings_page.dart';
-import 'package:weather/utils/border_text_style.dart';
+import 'package:weather/reusable_widgets/border_text_style.dart';
 import 'components/main_page_weather_content.dart';
 
 class HomePage extends StatefulWidget {
@@ -17,6 +16,11 @@ class HomePage extends StatefulWidget {
 
   @override
   State<HomePage> createState() => _HomePageState();
+}
+
+String splitBox(Box box) {
+  String boxToString = box.getAt(box.length - 1).cityName.toString();
+  return boxToString.split(',')[0];
 }
 
 class _HomePageState extends State<HomePage> {
@@ -48,19 +52,14 @@ class _HomePageState extends State<HomePage> {
         title: ValueListenableBuilder(
           valueListenable: Hive.box(favCity).listenable(),
           builder: (BuildContext context, value, Widget? child) {
-            var box = Hive.box(favCity);
-
+            Box box = Hive.box(favCity);
             return box.isEmpty
                 ? Text(
-                    'brak',
+                    'brak',style: TextStyle(color: Colors.blue),
                   )
                 : BorderTextStyle(
                     child: Text(
-                      box
-                          .getAt(box.length - 1)
-                          .cityName
-                          .toString()
-                          .split(',')[0],
+                      splitBox(box),
                       style: const TextStyle(
                         color: Colors.white,
                         fontSize: 30,
@@ -82,7 +81,7 @@ class _HomePageState extends State<HomePage> {
             decoration: IconDecoration(
               border: IconBorder(
                 color: Colors.black26,
-                width: 6,
+                width: 5,
               ),
             ),
           ),
